requirements.txt
absl-py==2.3.1
accelerate==1.9.0
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.17.2
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyio==4.10.0
args==0.1.0
asttokens==3.0.0
async-timeout==5.0.1
attrdict==2.0.1
attrs==25.3.0
audioread==3.0.1
banal==1.0.6
bitarray==3.6.1
blessed==1.21.0
boto3==1.40.0
botocore==1.40.0
braceexpand==0.1.7
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
clint==0.5.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
crcmod==1.7
cryptography==45.0.6
cycler==0.12.1
Cython==3.1.3
dataset==1.6.2
datasets==3.0.0
debugpy==1.8.16
decorator==5.2.1
deepseek-tokenizer==0.1.3
deepspeed==0.15.4
diffusers==0.35.1
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.0
fairscale==0.4.13
fairseq==0.12.2
filelock==3.18.0
flash_attn==2.6.1
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2024.2.0
ftfy==6.3.1
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
gpustat==1.1.1
greenlet==3.3.0
h11==0.16.0
hf-xet==1.1.5
hjson==3.1.0
hpsv2==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.3
hydra-core==1.0.7
icecream==2.1.7
idna==3.10
image-reward==1.5
importlib_metadata==8.7.0
iniconfig==2.1.0
ipykernel==6.30.1
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==0.10.0
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.10.1
llvmlite==0.44.0
lxml==6.0.1
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
modelscope==1.29.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
multiprocess==0.70.16
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
numba==0.61.2
numpy==2.2.6
nvidia-ml-py==12.575.51
omegaconf==2.0.6
open_clip_torch==3.1.0
openai==1.107.2
openai-clip==1.0.1
opencv-python==4.12.0.88
oss2==2.19.1
packaging==25.0
pandas==2.3.2
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
pooch==1.8.2
portalocker==3.2.0
prettytable==3.16.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyarrow-hotfix==0.7
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pyecharts==2.0.8
Pygments==2.19.2
pynvml==12.0.0
pyparsing==3.2.3
pytest==7.2.0
pytest-split==0.8.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.1
RapidFuzz==3.13.0
regex==2025.7.34
requests==2.32.4
rich==13.9.4
rouge-score==0.0.4
s3transfer==0.13.1
sacrebleu==2.5.1
safetensors==0.5.3
scipy==1.15.3
seaborn==0.13.2
sentencepiece==0.2.0
sentry-sdk==2.34.1
setproctitle==1.3.6
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==0.5.0.post1
SQLAlchemy==1.4.54
stack-data==0.6.3
swankit==0.2.4
swanlab==0.6.8
sympy==1.13.1
tabulate==0.9.0
tensorboardX==2.6.4
threadpoolctl==3.6.0
timm==1.0.19
tokenizers==0.21.4
tomli==2.2.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.0
trl==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.176
ultralytics-thop==2.0.15
unicodedata2==16.0.0
urllib3==2.5.0
wandb==0.18.3
wcwidth==0.2.13
webdataset==1.0.2
wrapt==1.17.2
xxhash==3.5.0
yapf==0.43.0
yarl==1.20.1
zhconv==1.4.3
zipp==3.23.0
